Add checking for badwords, increase recording time to 15s
Increased the recording time for the .wav files to 15s. Whisper can support up to 30s, but I feel this is a good compromise of both being responsive and accurate.
The issue with checking too frequently is that words can sometimes be cut-off. This can still happen, but there are less cut-off words the longer the recording.
Next up would be making a new file and a function within that file that executes when a bad word is detected. Almost there!!!